Support parsing less-nested sections after more-nested sections
Previously parsing a section of depth 1 after a section of depth 2 would
cause the parser to fail with an error. We now correctly parse trees of
sections.

The current implementation parses directly into a tree. This methodology
requires LL(k) parsing. Ideally, this would be replaced by a lexer and
parser combination to become a LL(1) parser.